ssl_stream_server/requirements.txt
# Certificate Transparency Log Monitoring
construct>=2.10.0
aiohttp>=3.9.0
PyOpenSSL>=23.0.0
python-dotenv>=1.0.0

# Database (PostgreSQL with async support)
psycopg[binary,pool]>=3.1.0
